[Core] Separate weex api result type from WeexApiHeader
