+++ ios/sdk/WeexSDK/Sources/Utility/WXVersion.m
@@ -20,8 +20,8 @@
 #import "WXVersion.h"
 #import "WXDefine.h"
 
-static const char* WeexSDKBuildTime = "2019-01-10 02:29:08 UTC";
-static const unsigned long WeexSDKBuildTimestamp = 1547087348;
+static const char* WeexSDKBuildTime = "2019-03-06 06:27:55 UTC";
+static const unsigned long WeexSDKBuildTimestamp = 1551853675;
 
 NSString* GetWeexSDKVersion(void)
 {
